requirements.txt
asgiref==3.10.0
boto3==1.40.74
botocore==1.40.74
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
Django==4.2.26
django-storages==1.14.6
djangorestframework==3.16.1
fonttools==4.60.1
gunicorn==21.2.0
idna==3.11
importlib_resources==6.5.2
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.4
narwhals==2.11.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
s3transfer==0.14.0
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
zipp==3.23.0
